Move node related to data_set_manager app
